Converted frequency to long primitive type (from integer) in anticipation of HackRF support which requires frequencies greater than 2.147 GHz, which is the max hertz value that can be represented by an integer primitive.

+++ trunk/src/controller/ConfigurationControllerModel.java
@@ -167,7 +167,7 @@ public class ConfigurationControllerModel extends DefaultTreeModel
      * Convenience method to add a new default system, site and channel to 
      * the tree
      */
-    public void createChannel( int frequency, DecoderType decoder )
+    public void createChannel( long frequency, DecoderType decoder )
     {
     	System system = new System();
