+++ trunk/src/decode/mpt1327/MPT1327CallEvent.java
@@ -25,14 +25,14 @@ import decode.DecoderType;
 public class MPT1327CallEvent extends CallEvent
 {
 	private int mChannel;
-	private int mFrequency;
+	private long mFrequency;
 	
 	public MPT1327CallEvent( CallEventType callEventType, 
 							 AliasList aliasList,
 							 String fromID,
 							 String toID,
 							 int channel,
-							 int frequency,
+							 long frequency,
 							 String details )
     {
 	    super( DecoderType.MPT1327, callEventType, 
@@ -71,7 +71,7 @@ public class MPT1327CallEvent extends CallEvent
     }
 
 	@Override
-    public int getFrequency()
+    public long getFrequency()
     {
 	    return mFrequency;
     }
@@ -87,7 +87,7 @@ public class MPT1327CallEvent extends CallEvent
 		private String mToID;
 		private String mDetails;
 		private int mChannel;
-		private int mFrequency;
+		private long mFrequency;
 
 		public Builder( CallEventType callEventType )
 		{
@@ -118,7 +118,7 @@ public class MPT1327CallEvent extends CallEvent
 			return this;
 		}
 		
-		public Builder frequency( int frequency )
+		public Builder frequency( long frequency )
 		{
 			mFrequency = frequency;
 			return this;
